# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /packages/v2/{id}:
    delete:
      tags:
      - Packages
      summary: ''
      description: 'Discontinues a Package at a specific Facility.</br><b>Permissions Required</b>: View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: delete_packages_v2_{id}
      operationId: DeletePackageById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to discontinue.
        format: ''
      responses:
        200:
          description: Ok
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a Package by its Id.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_{id}
      operationId: GetPackageById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Package will be validated against the specified License Number. If not specified, the Package will be validated against all of the User's current Facilities. Please note that if the Package is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/active:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of active packages for a specified Facility.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_active
      operationId: GetPackageActive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of active packages.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /packages/v2/adjust/reasons:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of adjustment reasons for packages at a specified Facility.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_packages_v2_adjust_reasons
      operationId: GetPackageAdjustReasons
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of adjustment reasons.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/{id}/source/harvests:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves the source harvests for a Package by its Id.</br><b>Permissions Required</b>: View Package Source Harvests'
      externalDocs:
        x-doc-key: get_packages_v2_{id}_source_harvests
      operationId: GetPackageByIdSourceHarvests
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Package will be validated against the specified License Number. If not specified, the Package will be validated against all of the User's current Facilities. Please note that if the Package is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/inactive:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of inactive packages for a specified Facility.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_inactive
      operationId: GetPackageInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of inactive packages.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /packages/v2/intransit:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of packages in transit for a specified Facility.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_intransit
      operationId: GetPackageIntransit
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of packages in transit.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /packages/v2/{label}:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a Package by its label.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_{label}
      operationId: GetPackageByLabel
      parameters:
      - type: string
        name: label
        in: path
        required: true
        description: Route parameter label
        format: ''
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Package will be validated against the specified License Number. If not specified, the Package will be validated against all of the User's current Facilities. Please note that if the Package is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/labsamples:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of lab sample packages created or sent for testing for a specified Facility.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_labsamples
      operationId: GetPackageLabSamples
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of lab sample packages that have been created/sent for testing.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /packages/v2/onhold:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of packages on hold for a specified Facility.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_onhold
      operationId: GetPackageOnHold
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of packages on hold.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /packages/v2/transferred:
    get:
      tags:
      - Packages
      summary: ''
      description: 'Retrieves a list of transferred packages for a specific Facility.</br><b>Permissions Required</b>: View Packages'
      externalDocs:
        x-doc-key: get_packages_v2_transferred
      operationId: GetPackageTransferred
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of transferred packages.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
...
  /packages/v2/:
    post:
      tags:
      - Packages
      summary: ''
      description: 'Creates new packages for a specified Facility.</br><b>Permissions Required</b>: View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: post_packages_v2_
      operationId: PostPackages
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record the list of new packages.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/adjust:
    post:
      tags:
      - Packages
      summary: ''
      description: 'Records a list of adjustments for packages at a specific Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: post_packages_v2_adjust
      operationId: PostPackageAdjust
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record the list of adjustments.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - Packages
      summary: ''
      description: 'Set the final quantity for a Package.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_adjust
      operationId: PutPackageAdjust
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record the list of adjustments.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/plantings:
    post:
      tags:
      - Packages
      summary: ''
      description: 'Creates new plantings from packages for a specified Facility.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants • View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: post_packages_v2_plantings
      operationId: PostPackagePlantings
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record the list of new plantings.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/testing:
    post:
      tags:
      - Packages
      summary: ''
      description: 'Creates new packages for testing for a specified Facility.</br><b>Permissions Required</b>: View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: post_packages_v2_testing
      operationId: PostPackageTesting
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record the list of new packages for testing.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/decontaminate:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates the Product decontaminate information for a list of packages at a specific Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_decontaminate
      operationId: PutPackageDecontaminate
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update the list of product decontaminations.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/donation/flag:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Flags one or more packages for donation at the specified Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_donation_flag
      operationId: PutPackageDonationFlag
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update list of flagged donations.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/donation/unflag:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Removes the donation flag from one or more packages at the specified Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_donation_unflag
      operationId: PutPackageDonationUnflag
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update list of unflaged donations.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/externalid:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates the external identifiers for one or more packages at the specified Facility.</br><b>Permissions Required</b>: View Packages • Manage Package Inventory • External Id Enabled'
      externalDocs:
        x-doc-key: put_packages_v2_externalid
      operationId: PutPackageExternalid
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update list of change external Ids.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/finish:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates a list of packages as finished for a specific Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_finish
      operationId: PutPackageFinish
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update the list of finish packages.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/item:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates the associated Item for one or more packages at the specified Facility.</br><b>Permissions Required</b>: View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: put_packages_v2_item
      operationId: PutPackageItem
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update list of changed items.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/labtests/required:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates the list of required lab test batches for one or more packages at the specified Facility.</br><b>Permissions Required</b>: View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: put_packages_v2_labtests_required
      operationId: PutPackageLabTestRequired
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update the list of required lab test batches.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/location:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates the Location and Sublocation for one or more packages at the specified Facility.</br><b>Permissions Required</b>: View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: put_packages_v2_location
      operationId: PutPackageLocation
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update the list of change locations.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/note:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates notes associated with one or more packages for the specified Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory • Manage Package Notes'
      externalDocs:
        x-doc-key: put_packages_v2_note
      operationId: PutPackageNote
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update list of change notes.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/remediate:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Updates a list of Product remediations for packages at a specific Facility.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_remediate
      operationId: PutPackageRemediate
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update the list of product remediations.
        format: ''
      responses:
        200:
          description: Ok
  /packages/v2/tradesample/flag:
    put:
      tags:
      - Packages
      summary: ''
      description: 'Flags or unflags one or more packages at the specified Facility as trade samples.</br><b>Permissions Required</b>: View Packages • Manage Packages Inventory'
      externalDocs:
        x-doc-key: put_packages_v2_tradesample_flag
      operationId: PutPackageTradeSampleFlag
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update trade sample flags.
        format: ''
      responses:
        200:
          description: Ok
...